+++ liveproxy/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '0.0.4.dev1'
+__version__ = '0.0.4.dev2'